Wconversion

+++ src/util/color.c
@@ -43,8 +43,8 @@ inline t_color
 {
 	uint8_t	c[3];
 
-	c[0] = ((x & 0xFF) * (y & 0xFF) + 128) >> 8;
-	c[1] = (((x >> 8) & 0xFF) * ((y >> 8) & 0xFF) + 128) >> 8;
-	c[2] = (((x >> 16) & 0xFF) * ((y >> 16) & 0xFF) + 128) >> 8;
+	c[0] = (uint8_t)(((x & 0xFF) * (y & 0xFF) + 128) >> 8);
+	c[1] = (uint8_t)((((x >> 8) & 0xFF) * ((y >> 8) & 0xFF) + 128) >> 8);
+	c[2] = (uint8_t)((((x >> 16) & 0xFF) * ((y >> 16) & 0xFF) + 128) >> 8);
 	return ((t_color)c[0] | ((t_color)c[1] << 8) | ((t_color)c[2] << 16));
 }
